extracts functionality to FlashcardDisplay component
